moved the cleaner to its own file, updated crawler, cleaner, and add-ons
